Menambahkan init pada saat server pertama kali dijalankan
