+++ app.js
@@ -3,7 +3,7 @@ const fastify = require("fastify")({
 });
 
 const { leaderboard, main } = require("./router");
-const { cacheAndConstant, commonSchema } = require("./plugin");
+const { cacheAndConstant, commonSchema, init } = require("./plugin");
 
 // plugin
 fastify.register(require("fastify-compression"));
@@ -16,5 +16,7 @@ const cachePreHandler = require("./common/cachePreHandler")(fastify);
 // routing
 fastify.register(main(cachePreHandler), { prefix: "/" });
 fastify.register(leaderboard(cachePreHandler), { prefix: "/leaderboard" });
+
+fastify.register(init);
 
 module.exports = fastify;
