+++ plugin/index.js
@@ -1,4 +1,5 @@
 const cacheAndConstant = require("./cacheAndConstant");
 const commonSchema = require("./commonSchema");
+const init = require("./init");
 
-module.exports = { cacheAndConstant, commonSchema };
+module.exports = { init, cacheAndConstant, commonSchema };
